pyproject.toml
[project]
name = "invisible-wroclaw"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dynaconf>=3.2.11",
    "peewee>=3.18.2",
    "python-telegram-bot>=22.4",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
]
